Move status icon to metric folder.
