+++ components/frontend/src/metric/StatusIcon.test.js
